requirements.txt
asgiref==3.2.3
colorama==0.4.3
Django==3.0.2
django-filter==2.2.0
django-nose==1.4.6
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
nose==1.3.7
pinocchio==0.4.2
psycopg2==2.8.4
Pygments==2.5.2
PyJWT==1.7.1
pytz==2019.3
rednose==1.3.0
sqlparse==0.3.0
termstyle==0.1.11
